Created a module for functions to normalize data (e.g., sanitizing the project name and swapping UUIDs for their names)
